requirements.txt
feedparser>=6.0.11
pyyaml>=6.0.2
requests>=2.32.0
anthropic>=0.30.0
openai>=1.50.0
google-genai>=1.0.0
